exporter: move restore logic here from snapshot
